test(core): migrate CustomTitleTests (Wave D)
33 tests, 71 failure sites preserved 1:1. The nine `return try expect(false, ...)` patterns inside rename-completion case destructures convert to Issue.record + return.
